lint error:
replace all ' = ' with '='

+++ ivy_tests/test_ivy/test_frontends/test_torch/test_indexing_slicing_joining_mutating_ops.py
@@ -74,19 +74,19 @@ def test_torch_cat(
     
 # reshape
 @given(
-    dtype_value_shape = helpers.dtype_and_values(
+    dtype_value_shape=helpers.dtype_and_values(
         available_dtypes=tuple(
             set(ivy_np.valid_float_dtypes).intersection(set(ivy_torch.valid_float_dtypes))
         ),
-        ret_shape = True,
+        ret_shape=True,
     ),
-    as_variable = st.booleans(),
-    with_out = st.booleans(),
-    num_positional_args = helpers.num_positional_args(
-        fn_name = "functional.frontends.torch.reshape"
+    as_variable=st.booleans(),
+    with_out=st.booleans(),
+    num_positional_args=helpers.num_positional_args(
+        fn_name="functional.frontends.torch.reshape"
     ),
-    native_array = st.booleans(),
-    copy = st.booleans(),
+    native_array=st.booleans(),
+    copy=st.booleans(),
 )
 def test_torch_reshape(
     dtype_value_shape,
@@ -97,18 +97,20 @@ def test_torch_reshape(
     fw,
     copy,
 ):
-    input_dtype, value, shape = dtype_value_shape
+    input_dtype, value, shape=dtype_value_shape
     helpers.test_frontend_function(
-        input_dtypes = input_dtype,
-        as_variable_flags = as_variable,
-        with_out = with_out,
-        num_positional_args = num_positional_args,
-        native_array_flags = native_array,
-        fw = fw,
-        frontend = "torch",
-        fn_name = "reshape",
-        input = np.asarray(value, dtype = input_dtype),
-        shape = shape,
-        copy = copy,
-        out = None,
+        input_dtypes=input_dtype,
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        fw=fw,
+        frontend="torch",
+        fn_name="reshape",
+        input=np.asarray(value, dtype=input_dtype),
+        shape=shape,
+        copy=copy,
+        out=None,
     )
+
+
